+++ packages/terra-doc-template/src/DocTemplate.jsx
@@ -161,7 +161,7 @@ const DocTemplate = ({ packageName, readme, srcPath, repositoryURL, branch, exam
 
       {localPropsTables.map(propsTable =>
         <PropsTable
-          src={propsTable.componentSource}
+          src={propsTable.componentSrc}
           componentName={propsTable.componentName}
           key={propsTable.id}
         />,