+++ packages/terra-doc-template/tests/jest/DocTemplate.test.jsx
@@ -47,14 +47,14 @@ describe('DocTemplate', () => {
   });
 
   it('should show one props table', () => {
-    const wrapper = shallow(<DocTemplate propsTables={[{ componentSource: testComponentSrc, componentName: 'Test Component' }]} />);
+    const wrapper = shallow(<DocTemplate propsTables={[{ componentSrc: testComponentSrc, componentName: 'Test Component' }]} />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should show multiple props tables', () => {
     const wrapper = shallow(<DocTemplate
-      propsTables={[{ componentSource: testComponentSrc, componentName: 'Test Component' },
-      { componentSource: testComponentSrc, componentName: 'Test Component (Again)' }]}
+      propsTables={[{ componentSrc: testComponentSrc, componentName: 'Test Component' },
+      { componentSrc: testComponentSrc, componentName: 'Test Component (Again)' }]}
     />);
     expect(wrapper).toMatchSnapshot();
   });
@@ -66,8 +66,8 @@ describe('DocTemplate', () => {
       repositoryURL="git+https://github.com/cerner/terra-core.git"
       examples={[{ title: 'Test Example 1', description: 'Describing the test', example: exampleElement, source: exampleSrc },
       { title: 'Test Example 2', description: 'Describing the test mk. 2', example: exampleElement, source: exampleSrc }]}
-      propsTables={[{ componentSource: testComponentSrc, componentName: 'Test Component' },
-      { componentSource: testComponentSrc, componentName: 'Test Component (Again)' }]}
+      propsTables={[{ componentSrc: testComponentSrc, componentName: 'Test Component' },
+      { componentSrc: testComponentSrc, componentName: 'Test Component (Again)' }]}
     />);
     expect(wrapper).toMatchSnapshot();
   });