+++ packages/terra-doc-template/tests/jest/DocTemplate.test.jsx
@@ -59,10 +59,11 @@ describe('DocTemplate', () => {
     expect(wrapper).toMatchSnapshot();
   });
 
-  it('should show the version, readme, examples, and props tables', () => {
+  it('should show the version, readme, source link, examples, and props tables', () => {
     const wrapper = shallow(<DocTemplate
       packageName="terra-doc-template"
       readme={readme}
+      srcPath="https://github.com/cerner/terra-core/tree/master/packages/terra-doc-template"
       examples={[{ title: 'Test Example 1', description: 'Describing the test', example: exampleElement, source: exampleSrc },
       { title: 'Test Example 2', description: 'Describing the test mk. 2', example: exampleElement, source: exampleSrc }]}
       propsTables={[{ componentSrc: testComponentSrc, componentName: 'Test Component' },