+++ packages/terra-doc-template/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Add syntax highlighting to source code examples
 
 1.0.0 - (May 2, 2018)
 ------------------