+++ packages/terra-doc-template/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Add syntax highlighting to source code examples
+
 ### Fixed
 * Update hash key reference in render to match the defined key in the propsTables prop