+++ packages/terra-doc-template/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Changed
 * Add syntax highlighting to source code examples
 
+### Fixed
+* Update hash key reference in render to match the defined key in the propsTables prop
+
 1.0.0 - (May 2, 2018)
 ------------------
 * Initial stable release