+++ packages/terra-doc-template/CHANGELOG.md
@@ -3,6 +3,11 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Add syntax highlighting to source code examples
+
+### Fixed
+* Update hash key reference in render to match the defined key in the propsTables prop
 
 1.0.0 - (May 2, 2018)
 ------------------