+++ packages/terra-doc-template/CHANGELOG.md
@@ -3,8 +3,12 @@ ChangeLog
 
 Unreleased
 ----------
-### Added
-* Added a link to the component's source code
+### Changed
+* Add syntax highlighting to source code examples
+* Add a link to the component's source code
+
+### Fixed
+* Update hash key reference in render to match the defined key in the propsTables prop
 
 1.0.0 - (May 2, 2018)
 ------------------