+++ packages/terra-doc-template/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Add syntax highlighting to source code examples
+* Add a link to the component's source code
 
 ### Fixed
 * Update hash key reference in render to match the defined key in the propsTables prop