+++ packages/terra-doc-template/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+1.1.0 - (May 8, 2018)
+------------------
 ### Changed
 * Add syntax highlighting to source code examples
 * Add a link to the component's source code